fixes #1840
